サインアップとログインのコンポーネントをリファクタリング
