+++ src/app/(public)/signup/_components/SignupPage.tsx
@@ -1,0 +1,173 @@
+"use client";
+
+import NextLink from "next/link";
+import { useState, useCallback, useEffect } from "react";
+import { createSupabaseBrowserClient } from "@/app/_libs/supabase/browserClient";
+import { zodResolver } from "@hookform/resolvers/zod";
+import { useForm, FormProvider } from "react-hook-form";
+
+import { Button } from "@/app/_components/ui/button";
+import { FormTextField } from "@/app/_components/FormTextField";
+import { FormErrorMessage } from "@/app/_components/FormErrorMessage";
+
+import { LuSend } from "react-icons/lu";
+import type { SignupRequest } from "@/app/_types/SignupRequest";
+import { signupRequestSchema } from "@/app/_types/SignupRequest";
+import { appBaseUrl } from "@/app/_configs/app-config";
+
+const c_Email = "email";
+const c_Password = "password";
+const c_ConfirmPassword = "confirmPassword";
+
+export const SignupPage: React.FC = () => {
+  const [isSignUpCompleted, setIsSignUpCompleted] = useState(false);
+
+  const form = useForm<SignupRequest>({
+    mode: "onChange",
+    resolver: zodResolver(signupRequestSchema),
+  });
+  const fieldErrors = form.formState.errors;
+
+  // ルートエラーのクリア
+  const clearRootError = useCallback(() => {
+    if (fieldErrors.root) form.clearErrors("root");
+  }, [fieldErrors.root, form]);
+
+  // サーバサイドで発生した問題をルートエラーとして設定
+  const setRootError = useCallback(
+    (errorMsg: string) => {
+      form.setError("root", {
+        type: "manual",
+        message: errorMsg,
+      });
+    },
+    [form],
+  );
+
+  // confirmPassword が先に入力され、その後 password が変更された場合、
+  // zod の refine() による一致チェックが再評価されず、バリデーションエラーが反映されない問題に対応
+  // password の変更時に confirmPassword のバリデーションを明示的に再実行するようにしている
+  const password = form.watch(c_Password);
+  useEffect(() => {
+    form.trigger(c_ConfirmPassword);
+  }, [password, form]);
+
+  // サインアップフォーム送信の処理 Server Action（Custom Invocation）で処理
+  const onSubmit = useCallback(
+    async (formValues: SignupRequest) => {
+      const supabase = createSupabaseBrowserClient();
+
+      // Supabase で AutoConfirm が ON の場合、signUp () 完了時点で
+      // ログイン状態になることに注意。
+      const { error } = await supabase.auth.signUp({
+        email: formValues.email,
+        password: formValues.password,
+        options: {
+          emailRedirectTo: `${appBaseUrl}/login`,
+        },
+      });
+      if (error) {
+        switch (error.code) {
+          case "over_email_send_rate_limit":
+            setRootError(
+              "システムから送信可能な認証メールが規定数を超えました。しばらくたってから、再度、試してみてください。",
+            );
+            break;
+          case "user_already_exists":
+            setRootError(
+              `メールアドレス ( ${formValues.email} ) は、既に使用されています。`,
+            );
+            break;
+          case "weak_password":
+            setRootError("もう少し安全性の高いパスワードを設定してください。");
+            break;
+          default:
+            setRootError(`サインアップ処理に失敗しました ( ${error.code} )`);
+            break;
+        }
+        return;
+      }
+
+      // ローカル環境の開発用 Supabase では、サインアップで自動ログインになるため、
+      // 一旦、ログアウト処理をしておく。本番環境では以下のログアウト処理は不要。
+      await supabase.auth.signOut();
+
+      setIsSignUpCompleted(true);
+    },
+    [setRootError],
+  );
+
+  return (
+    <div className="flex justify-center pt-12">
+      <div className="w-full max-w-[460px]">
+        <h1 className="mb-8 text-center text-3xl font-bold">サインアップ</h1>
+        <form
+          className="space-y-4"
+          noValidate
+          onSubmit={form.handleSubmit(onSubmit)}
+        >
+          <FormProvider {...form}>
+            <FormTextField<SignupRequest>
+              fieldKey={c_Email}
+              labelText="メールアドレス"
+              placeholder="name@example.com"
+              autoComplete="email"
+              registerOnChange={clearRootError}
+            />
+
+            <FormTextField<SignupRequest>
+              fieldKey={c_Password}
+              labelText="パスワード"
+              type="password"
+              placeholder="password"
+              registerOnChange={clearRootError}
+            />
+
+            <FormTextField<SignupRequest>
+              fieldKey={c_ConfirmPassword}
+              labelText="パスワード（確認用）"
+              type="password"
+              placeholder="password"
+              registerOnChange={clearRootError}
+            />
+          </FormProvider>
+
+          <FormErrorMessage msg={fieldErrors.root?.message} />
+
+          <div>
+            <Button
+              className="w-full"
+              type="submit"
+              disabled={!form.formState.isValid}
+            >
+              <div className="flex items-center gap-x-1">
+                <LuSend />
+                <div>登録</div>
+              </div>
+            </Button>
+          </div>
+        </form>
+        {isSignUpCompleted && (
+          <div className="mt-4">
+            サインアップが完了しました。
+            {/* 登録いただいたメールアドレス ( {form.getValues(c_Email)} ) 宛に、認証メールを送信しました。メールに記載のURLをクリックして、登録手続きを完了してください。 */}
+            <NextLink
+              href={`/login?${c_Email}=${form.getValues(c_Email)}`}
+              className="px-0.5 text-blue-500 underline"
+            >
+              ログインページ
+            </NextLink>
+            からログインしてください。
+          </div>
+        )}
+        <div className="mt-2 text-sm">
+          既にアカウントをお持ちの方は
+          <NextLink href="/login" className="px-0.5 text-blue-500 underline">
+            ログインページ
+          </NextLink>
+          をご利用ください。
+        </div>
+      </div>
+    </div>
+  );
+};
