+++ src/app/_libs/authenticateUser.ts
@@ -4,6 +4,12 @@ import { SupabaseUserNotFoundError } from "@/app/_libs/errors";
 import { UserService } from "@/app/_services/userService";
 import prisma from "@/app/_libs/prisma";
 
+const getSupabaseUserOrNull = async (): Promise<User | null> => {
+  const supabase = await createSupabaseServerClient();
+  const { data, error } = await supabase.auth.getUser();
+  return error || !data?.user ? null : data.user;
+};
+
 /**
  * Supabaseのセッション情報をもとに、認証済みのユーザー情報を取得する関数。
  *
@@ -14,13 +20,20 @@ import prisma from "@/app/_libs/prisma";
  * @throws SupabaseUserNotFoundError - ユーザーが存在しない、もしくはセッションが無効な場合
  */
 export const authenticateSupabaseUser = async (): Promise<User> => {
-  const supabase = await createSupabaseServerClient();
-  const { data, error } = await supabase.auth.getUser();
-  if (error || !data?.user) {
-    throw new SupabaseUserNotFoundError();
-  }
-  return data.user;
+  const user = await getSupabaseUserOrNull();
+  if (!user) throw new SupabaseUserNotFoundError();
+  return user;
 };
+
+/**
+ * Supabaseのセッション情報をもとに、認証済みのユーザー情報を取得する関数（存在しない場合は null を返す）。
+ *
+ * この関数はサーバー上でSupabaseの認証セッションを確認し、ユーザー情報を取得します。
+ * セッションが無効、またはユーザー情報が取得できない場合は null を返します。
+ *
+ * @returns 認証済みSupabaseユーザー情報（User オブジェクト）、または null
+ */
+export const maybeAuthenticateSupabaseUser = getSupabaseUserOrNull;
 
 /**
  * Supabase上の認証に成功したユーザーに対応する、アプリケーション側のユーザーデータを取得する関数。
